requirements.txt
numpy
pandas
sklearn.processing
LabelEncoder
io
BytesIO